created MenuStateContainer
App Routers route to MenuStateContainer.

Conditional rendering for Menu is at MenuStateContainer
